+++ src/main/java/com/github/trilonka/subscribemailertelegrambot/client/dto/LikesInfo.java
@@ -1,0 +1,10 @@
+package com.github.trilonka.subscribemailertelegrambot.client.dto;
+
+import lombok.Data;
+
+@Data
+public class LikesInfo {
+
+    private Integer count;
+    private LikeStatus status;
+}