+++ src/main/java/com/github/trilonka/subscribemailertelegrambot/client/dto/PostType.java
@@ -1,0 +1,6 @@
+package com.github.trilonka.subscribemailertelegrambot.client.dto;
+
+public enum PostType {
+
+    UNKNOWN, USUAL, INNER_LINK, OUTER_LINK
+}