+++ src/main/java/com/github/trilonka/subscribemailertelegrambot/client/dto/VisibilityStatus.java
@@ -1,0 +1,12 @@
+package com.github.trilonka.subscribemailertelegrambot.client.dto;
+
+public enum VisibilityStatus {
+
+    UNKNOWN,
+    RESTRICTED,
+    PUBLIC,
+    PROTECTED,
+    PRIVATE,
+    DISABLED,
+    DELETED
+}