+++ src/main/java/com/github/trilonka/subscribemailertelegrambot/service/GroupSubService.java
@@ -3,6 +3,7 @@ package com.github.trilonka.subscribemailertelegrambot.service;
 import com.github.trilonka.subscribemailertelegrambot.client.dto.GroupDiscussionInfo;
 import com.github.trilonka.subscribemailertelegrambot.repository.entity.GroupSub;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface GroupSubService {
@@ -12,4 +13,6 @@ public interface GroupSubService {
     GroupSub save(GroupSub groupSub);
 
     Optional<GroupSub> findById(Integer id);
+
+    List<GroupSub> findAll();
 }