+++ src/main/java/com/github/trilonka/subscribemailertelegrambot/service/impl/SendBotMessageServiceBasic.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+import java.util.List;
+
 
 @Service
 public class SendBotMessageServiceBasic implements SendBotMessageService {
@@ -29,5 +31,12 @@ public class SendBotMessageServiceBasic implements SendBotMessageService {
             // TODO: add logging to the project
             e.printStackTrace();
         }
+    }
+
+    @Override
+    public void sendMessage(String chatId, List<String> messages) {
+        if (messages.isEmpty()) return;
+
+        messages.forEach(it -> sendMessage(chatId, it));
     }
 }